wc: major improvements in disease panel components
